pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-rag"
version = "0.1.0"
description = "A standalone Agent RAG component with Chat and Deep Research capabilities"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Onyx Team" }
]
keywords = ["rag", "agent", "llm", "search", "vespa", "deep-research"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "litellm>=1.40.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "tiktoken>=0.5.0",
    "numpy>=1.24.0",
    "jinja2>=3.0.0",
    "qdrant-client>=1.9.0",
]

[project.optional-dependencies]
database = [
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "greenlet>=3.0.0",
]
vespa = [
    "pyvespa>=0.45.0",
]
parsing = [
    # Document parsers
    "pypdf>=3.0.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.0",
    # Fallback to Unstructured API
    "unstructured-client>=0.20.0",
]
chunking = [
    # Semantic chunking
    "chonkie>=0.1.0",
]
api = [
    # FastAPI ecosystem
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "python-multipart>=0.0.6",
]
ingestion = [
    "agent-rag[database,parsing,chunking,api]",
    "minio>=7.2.0",
    "hatchet-sdk>=0.30.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = [
    "agent-rag[database,vespa,ingestion,dev]",
]

[project.scripts]
agent-rag-init = "agent_rag.scripts.init_db:main"
agent-rag-migrate = "agent_rag.scripts.migrate:main"
agent-rag-vespa = "agent_rag.scripts.deploy_vespa:main"

[project.urls]
Homepage = "https://github.com/onyx-dot-app/onyx"
Documentation = "https://docs.onyx.app"
Repository = "https://github.com/onyx-dot-app/onyx"

[tool.hatch.build.targets.wheel]
packages = ["agent_rag"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "deepeval>=3.9.2",
]
